Moves wait_status and related code to their own files
